backend/requirements/requirements-noble.txt
acme==2.9.0
cryptography==41.0.7
fpdf2==2.7.8
h2==4.1.0
idna==3.6
priority==2.0.0
pyopenSSL==23.2.0
pynacl==1.5.0
pyotp==2.9.0
python_gnupg==0.5.2
sqlalchemy==1.4.50
twisted==24.3.0
txtorcon==23.11.0
